Refactoring: Use MEF to modularize ClearMine.Logic.dll.
git-svn-id: http://hugogu/svn/main/Workspace/Projects/TechnicalResearch/WPF/ClearMine@83 abc3d51d-b5e4-b64c-aea7-7d3d034b83e7

+++ ClearMine.Common/HeroHistory.cs
@@ -14,8 +14,11 @@
     [XmlRoot("record")]
     public class HistoryRecord 
     {
+        /// <summary>
+        /// The time used in seconds.
+        /// </summary>
         [XmlAttribute("score")]
-        public int Score { get; set; }
+        public double Score { get; set; }
         [XmlAttribute("when")]
         public DateTime Date { get; set; }
         [XmlAttribute("shoot")]
@@ -98,7 +101,7 @@
             }
         }
 
-        public void IncreaseWon(int score, DateTime time, string screenShoot)
+        public void IncreaseWon(double score, DateTime time, string screenShoot)
         {
             AverageScore = (GameWon * AverageScore + score) / ++GameWon;
             ++GamePlayed;
